@@ -3,11 +3,11 @@ package com.tbuonomo.viewpagerdotsindicator
 import android.graphics.drawable.GradientDrawable
 
 class DotsGradientDrawable : GradientDrawable() {
-  var currentColor: Int = 0
-    private set
+    var currentColor: Int = 0
+        private set
 
-  override fun setColor(argb: Int) {
-    super.setColor(argb)
-    currentColor = argb
-  }
+    override fun setColor(argb: Int) {
+        super.setColor(argb)
+        currentColor = argb
+    }
 }
